requirements.txt
flask==2.3.0
flask-cors==4.0.0
sentence-transformers==2.2.2
pandas==2.0.0
openpyxl==3.1.0
gspread==5.9.0
google-auth==2.22.0
requests==2.31.0
python-dotenv==1.0.0
sendgrid==6.10.0
scikit-learn==1.3.0
pytest==7.3.0
pytest-mock==3.10.0
pytest-asyncio==0.21.0
responses==0.23.0
pyyaml==6.0
